split systemd-journal further
